chore: 修复构建脚本移动文件夹命令错误
* 修正 git pull 命令

* home.vue文件添加console信息

+++ deployed.sh
@@ -3,9 +3,15 @@ USER='www'
 USER_GROUP='www'
 
 cd /home/git/vue-cli/
-git pull origin master
+git fetch --all
+git reset --hard origin/master
 sudo npm install
 npm run build
+if [ $? -eq 0 ]
+then
 chown -R $USER:$USER_GROUP $SITE_PATH
 rm -rf /home/projects/web/dist/*
-mv /home/git/vue-cli/dist/* /home/projects/web/
+cp -R /home/git/vue-cli/dist/ /home/projects/web/ -f
+else
+ echo 'npm run build failed'
+fi
